knowledge_base_rag/pyproject.toml
[project]
name = "knowledge-base-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Konrad Jarzyna",email = "konrad.jarzyna0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "chromadb (==0.5.12)",
    "langchain-chroma (>=0.1.2)",
    "google-adk (>=0.5.0,<0.6.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "a2a-sdk (>=0.2.5,<0.3.0)",
    "uvicorn (>=0.34.3,<0.35.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.packages]]
include = "src"

[tool.poetry.scripts]
start-web = "cli:start_web"
start-api = "cli:start_api"
start-a2a = "cli:start_a2a"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
black = "^25.1.0"
isort = "^6.0.1"
